+++ src/components/DropdownMenu/DropdownMenu.types.ts
@@ -1,5 +1,6 @@
 import type React from "react";
 import type { PopoverProps, PopoverInstance } from "components/Popover";
+import type { MenuItemProps } from "components/MenuItem";
 import type { FlyoutContentProps } from "components/_private/Flyout";
 
 export type Instance = PopoverInstance;
@@ -29,6 +30,8 @@ export type Props = Pick<
 
 export type ContentProps = Pick<FlyoutContentProps, "attributes" | "children" | "className">;
 
+export type ItemProps = Omit<MenuItemProps, "roundedCorners">;
+
 export type SectionProps = {
 	children: React.ReactNode;
 };
@@ -37,6 +40,4 @@ export type SubMenuProps = {
 	children: React.ReactNode;
 };
 
-export type SubTriggerProps = {
-	children: React.ReactNode;
-};
+export type SubTriggerProps = Omit<MenuItemProps, "endSlot" | "roundedCorners">;