+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 - trapFocus: Added contenteditable support + elements with custom keydown events
 - trapFocus: Added correct handling for focusable html tags with tabIndex=-1
 - Flyout: Updated outside click handlers to only trigger when it's active
+- DropdownMenu.SubTrigger: Added all DropdownMenu.Item props support
 
 Documentation